requirements.txt
amqp==5.3.1
asgiref==3.9.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.5
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
requests==2.32.3 
